Fix 64 to 32 bit narrowing in dawn::wire
Bug: dawn:1377
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/87671
Commit-Queue: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/dawn/wire/server/ServerAdapter.cpp
@@ -93,7 +93,7 @@ namespace dawn::wire::server {
             }
         }
 
-        cmd.featuresCount = features.size();
+        cmd.featuresCount = static_cast<uint32_t>(features.size());
         cmd.features = features.data();
 
         WGPUSupportedLimits limits = {};
